Writing the matmul using AVX2 intrinsics shaves about 10% off of the time. FMA doesn't speed things up relative to AVX multiply and add, but I use it anyway for now.

```
(eel) eric@zinnia:~/development/eel$ ./scripts/main.py data/Qwen2.5-0.5B-Instruct/
Prompt time: 1.9780s
Total time: 5.2664s
```
vs
```
(eel) eric@zinnia:~/development/eel$ ./scripts/main.py data/Qwen2.5-0.5B-Instruct/
Prompt time: 2.2554s
Total time: 5.9787s
```